pyproject.toml
[project]
name = "gen-ai-framework"
version = "0.1.0"
description = "Modular Gen AI Framework with separate API, LLM, RAG components and domain clients"
requires-python = ">=3.10"
dependencies = [
    "fastapi>=0.104.0",
    "uvicorn[standard]>=0.24.0",
    "pydantic>=2.5.0",
    "pydantic-settings>=2.1.0",
    "python-multipart>=0.0.6",
    "python-dotenv>=1.0.0",
    "langchain>=0.1.0",
    "langchain-openai>=0.0.2",
    "langchain-community>=0.0.10",
    "langchain-google-genai>=2.0.0",
    "openai>=1.6.0",
    "chromadb>=0.4.18",
    "pypdf>=3.17.0",
    "python-docx>=1.1.0",
    "openpyxl>=3.1.2",
    "aiofiles>=23.2.1",
    "httpx>=0.25.2",
    # MCP
    "mcp>=1.0.0",
    # OCR
    "easyocr>=1.7.0",
    "Pillow>=10.0.0",
    # Docling
    "docling>=2.0.0",
    # PDF with OCR fallback (PyMuPDF + pytesseract)
    "pymupdf>=1.23.0",
    "pytesseract>=0.3.10",
    "opencv-python-headless>=4.8.0",
    "numpy>=1.24.0",
    # Validations (fuzzy match)
    "rapidfuzz>=3.0.0",
    # LangChain doc loaders (PyPDFLoader, etc. in langchain_community)
    # Embeddings (optional local models)
    "sentence-transformers>=2.2.0",
    # RAG: hybrid search (BM25)
    "rank_bm25>=0.2.2",
    # Task queue (optional: install with pip install gen-ai-framework[celery])
    # LangGraph
    "langgraph>=0.2.0",
    "langgraph-checkpoint>=2.0.0",
]

[project.optional-dependencies]
celery = ["celery[redis]>=5.3.0", "redis>=5.0.0"]
vectorstore-pinecone = ["pinecone-client>=3.0.0"]
vectorstore-weaviate = ["weaviate-client>=4.0.0"]
vectorstore-qdrant = ["qdrant-client>=1.7.0"]
vectorstore-pgvector = ["psycopg2-binary>=2.9.0", "pgvector>=0.2.0"]
dev = ["pytest>=7.4.3", "pytest-asyncio>=0.21.1", "black>=23.11.0", "ruff>=0.1.6"]

[project.scripts]
batch-run = "src.clients.batch.runner:main"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.black]
line-length = 100
target-version = ['py310']

[tool.ruff]
line-length = 100
target-version = "py310"

[tool.pytest.ini_options]
testpaths = ["tests"]
pythonpath = ["src"]
python_files = ["test_*.py"]
python_functions = ["test_*"]
